refactor(stencils): upgrade a stencil in one atomic transaction
UpdateStencilInTemplate created the draft by dispatching CreateVersion
through the mediator (its own transaction) and then upgraded it in a second
transaction — non-atomic, and it left a stray draft if the upgrade failed.

Extracted CreateVersion's draft-creation into a Handle-scoped DraftVersionFactory
that both CreateVersion and UpdateStencilInTemplate call inside their own
transaction, so draft-create + stencil-upgrade now commit (or roll back)
together as one transaction. Behaviour is unchanged; the draft model is read
straight off the returned TemplateVersion instead of re-querying.
